+++ src/Plugin/Wechat/V3/Marketing/Transfer/CancelPlugin.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Plugin\Wechat\V3\Marketing\MchTransfer;
+namespace Yansongda\Pay\Plugin\Wechat\V3\Marketing\Transfer;
 
 use Closure;
 use Yansongda\Artful\Contract\PluginInterface;
@@ -17,9 +17,9 @@ use Yansongda\Pay\Pay;
 use function Yansongda\Pay\get_provider_config;
 
 /**
- * @see https://pay.weixin.qq.com/doc/v3/merchant/4012716437
+ * @see https://pay.weixin.qq.com/doc/v3/merchant/4012716458
  */
-class QueryPlugin implements PluginInterface
+class CancelPlugin implements PluginInterface
 {
     /**
      * @throws InvalidParamsException
@@ -28,26 +28,26 @@ class QueryPlugin implements PluginInterface
      */
     public function assembly(Rocket $rocket, Closure $next): Rocket
     {
-        Logger::debug('[Wechat][Marketing][MchTransfer][QueryPlugin] 插件开始装载', ['rocket' => $rocket]);
+        Logger::debug('[Wechat][Marketing][Transfer][CancelPlugin] 插件开始装载', ['rocket' => $rocket]);
 
         $config = get_provider_config('wechat', $rocket->getParams());
         $payload = $rocket->getPayload();
         $outBillNo = $payload?->get('out_bill_no') ?? null;
 
         if (Pay::MODE_SERVICE === ($config['mode'] ?? Pay::MODE_NORMAL)) {
-            throw new InvalidParamsException(Exception::PARAMS_PLUGIN_ONLY_SUPPORT_NORMAL_MODE, '参数异常: 通过商户单号查询转账单，只支持普通商户模式，当前配置为服务商模式');
+            throw new InvalidParamsException(Exception::PARAMS_PLUGIN_ONLY_SUPPORT_NORMAL_MODE, '参数异常: 撤销转账，只支持普通商户模式，当前配置为服务商模式');
         }
 
         if (empty($outBillNo)) {
-            throw new InvalidParamsException(Exception::PARAMS_NECESSARY_PARAMS_MISSING, '参数异常: 通过商户单号查询转账单，参数缺少 `out_bill_no`');
+            throw new InvalidParamsException(Exception::PARAMS_NECESSARY_PARAMS_MISSING, '参数异常: 撤销转账，参数缺少 `out_bill_no`');
         }
 
         $rocket->setPayload([
-            '_method' => 'GET',
-            '_url' => 'v3/fund-app/mch-transfer/transfer-bills/out-bill-no/'.$outBillNo,
+            '_method' => 'POST',
+            '_url' => 'v3/fund-app/mch-transfer/transfer-bills/out-bill-no/'.$outBillNo.'/cancel',
         ]);
 
-        Logger::info('[Wechat][Marketing][MchTransfer][QueryPlugin] 插件装载完毕', ['rocket' => $rocket]);
+        Logger::info('[Wechat][Marketing][Transfer][CancelPlugin] 插件装载完毕', ['rocket' => $rocket]);
 
         return $next($rocket);
     }