+++ tests/Plugin/Wechat/V3/Marketing/Transfer/CreatePluginTest.php
@@ -9,8 +9,8 @@ use Yansongda\Artful\Contract\HttpClientInterface;
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Artful\Exception\InvalidParamsException;
 use Yansongda\Pay\Pay;
-use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Transfer\CreatePlugin;
 use Yansongda\Artful\Rocket;
+use Yansongda\Pay\Plugin\Wechat\V3\Marketing\Transfer\CreatePlugin;
 use Yansongda\Pay\Tests\TestCase;
 use Yansongda\Supports\Collection;
 
@@ -60,9 +60,10 @@ class CreatePluginTest extends TestCase
 
         self::assertEquals([
             '_method' => 'POST',
-            '_url' => 'v3/transfer/batches',
+            '_url' => 'v3/fund-app/mch-transfer/transfer-bills',
             'test' => 'yansongda',
             'appid' => '1111',
+            'notify_url' => 'https://pay.yansongda.cn',
         ], $result->getPayload()->all());
     }
 
@@ -77,9 +78,10 @@ class CreatePluginTest extends TestCase
 
         self::assertEquals([
             '_method' => 'POST',
-            '_url' => 'v3/transfer/batches',
+            '_url' => 'v3/fund-app/mch-transfer/transfer-bills',
             'test' => 'yansongda',
             'appid' => 'wx55955316af4ef13',
+            'notify_url' => 'https://pay.yansongda.cn',
         ], $result->getPayload()->all());
     }
 
@@ -88,12 +90,8 @@ class CreatePluginTest extends TestCase
         $rocket = new Rocket();
         $rocket->setPayload(new Collection([
             "test" => "111",
-            'transfer_detail_list' => [
-                [
-                    'user_name' => 'yansongda',
-                    'foo' => 'bar',
-                ]
-            ]
+            'user_name' => 'yansongda',
+            'foo' => 'bar',
         ]));
 
         $result = $this->plugin->assembly($rocket, function ($rocket) {
@@ -102,13 +100,14 @@ class CreatePluginTest extends TestCase
 
         $payload = $result->getPayload()->all();
         self::assertEquals('POST', $payload['_method']);
-        self::assertEquals('v3/transfer/batches', $payload['_url']);
+        self::assertEquals('v3/fund-app/mch-transfer/transfer-bills', $payload['_url']);
         self::assertEquals('wx55955316af4ef13', $payload['appid']);
+        self::assertEquals('https://pay.yansongda.cn', $payload['notify_url']);
         self::assertEquals('111', $payload['test']);
+        self::assertEquals('bar', $payload['foo']);
         self::assertArrayHasKey('_serial_no', $payload);
-        self::assertArrayHasKey('user_name', $payload['transfer_detail_list'][0]);
-        self::assertNotEquals('yansongda', $payload['transfer_detail_list'][0]['user_name']);
-        self::assertEquals('bar', $payload['transfer_detail_list'][0]['foo']);
+        self::assertArrayHasKey('user_name', $payload);
+        self::assertNotEquals('yansongda', $payload['user_name']);
     }
 
     public function testNormalWithNameEmptyWechatCer()
@@ -116,12 +115,8 @@ class CreatePluginTest extends TestCase
         $rocket = new Rocket();
         $rocket->setParams(['_config' => 'empty_wechat_public_cert'])->setPayload(new Collection([
             "test" => "111",
-            'transfer_detail_list' => [
-                [
-                    'user_name' => 'yansongda',
-                    'foo' => 'bar',
-                ]
-            ]
+            'user_name' => 'yansongda',
+            'foo' => 'bar',
         ]));
 
         $response = new Response(
@@ -155,12 +150,32 @@ class CreatePluginTest extends TestCase
 
         $payload = $result->getPayload()->all();
         self::assertEquals('POST', $payload['_method']);
-        self::assertEquals('v3/transfer/batches', $payload['_url']);
+        self::assertEquals('v3/fund-app/mch-transfer/transfer-bills', $payload['_url']);
         self::assertEquals('wx55955316af4ef13', $payload['appid']);
         self::assertEquals('111', $payload['test']);
+        self::assertEquals('bar', $payload['foo']);
         self::assertArrayHasKey('_serial_no', $payload);
-        self::assertArrayHasKey('user_name', $payload['transfer_detail_list'][0]);
-        self::assertNotEquals('yansongda', $payload['transfer_detail_list'][0]['user_name']);
-        self::assertEquals('bar', $payload['transfer_detail_list'][0]['foo']);
+        self::assertArrayHasKey('user_name', $payload);
+        self::assertNotEquals('yansongda', $payload['user_name']);
+    }
+
+    public function testNormalParamsWithNotifyUrl()
+    {
+        $rocket = new Rocket();
+        $rocket->setPayload(new Collection( [
+            "test" => "yansongda",
+            'appid' => '1111',
+            'notify_url' => 'https://yansongda.cn',
+        ]));
+
+        $result = $this->plugin->assembly($rocket, function ($rocket) { return $rocket; });
+
+        self::assertEquals([
+            '_method' => 'POST',
+            '_url' => 'v3/fund-app/mch-transfer/transfer-bills',
+            'test' => 'yansongda',
+            'appid' => '1111',
+            'notify_url' => 'https://yansongda.cn',
+        ], $result->getPayload()->all());
     }
 }